implemented account deletion, specific error handling for all routes

+++ models/user.js
@@ -1,6 +1,8 @@
 const db = require("../db");
 const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
+const Pet = require('./pet');
+const Pound = require('./pound');
 
 
 /** Collection of related methods for users. */
@@ -34,12 +36,12 @@ class User {
 
     // Register a new user
     static async register({ username, email, password }) {
-        let existingUser = await this.find({ email });
+        let existingUser = await this.find({ username });
+        if (existingUser) throw new Error('Username already exists');
+
+        existingUser = await this.find({ email });
         if (existingUser) throw new Error('Email already exists');
 
-        existingUser = await this.find({ username });
-        if (existingUser) throw new Error('Username already exists');
-    
         const hashedPassword = await bcrypt.hash(password, 10);
     
         try {
@@ -60,7 +62,10 @@ class User {
     static async login({ email, password }) {
         const user = await this.find({ email });
         // console.log(user) // sanity check
-        if (!user || !(await bcrypt.compare(password, user.password))) throw new Error('Invalid credentials');
+        if (!user || !(await bcrypt.compare(password, user.password))) {
+            throw new Error('Invalid credentials');
+        }
+
         try {
             const token = jwt.sign({ userId: user.id }, process.env.JWT_SECRET, { expiresIn: '1h' });
             return { token, user: { id: user.id, username: user.username, email: user.email } };
@@ -94,13 +99,31 @@ class User {
     }
     
     // Delete user
-    static async delete(userId) {
+    static async delete(userId, { username, email, password }) { 
         try {
+            const user = await this.find({ id: userId });
+    
+            if (!user || user.username !== username || user.email !== email || !(await bcrypt.compare(password, user.password))) {
+                throw new Error('Invalid credentials');
+            }
+
+            const userPets = await Pet.findByOwnerId(userId);
+            console.log(userPets)
             const result = await db.query('DELETE FROM users WHERE id = $1 RETURNING id;', [userId]);
+            if (result.rowCount > 0) {
+                for (const pet of userPets) {
+                    await Pound.abandon(pet.id); 
+                }
+            }
+    
             return result.rowCount > 0;
         } catch (error) {
-            console.error('Error deleting user:', error);
-            throw new Error('Deletion failed');
+            console.error('Error deleting user:', error.stack);
+            if (error.message === 'Invalid credentials') {
+                throw error; 
+            } else {
+                throw new Error('Deletion failed');
+            }
         }
     }
 }
